Move tests to subfolders and pass in addressmanager address
