+++ test/libs/byte_utils.js
