+++ test/oracles/oracle_factory.js
@@ -1,8 +1,8 @@
 const AddressManager = artifacts.require("./storage/AddressManager.sol");
 const OracleFactory = artifacts.require('./oracles/OracleFactory.sol');
 const Oracle = artifacts.require('./oracles/Oracle.sol');
-const BlockHeightManager = require('./helpers/block_height_manager');
-const Utils = require('./helpers/utils');
+const BlockHeightManager = require('../helpers/block_height_manager');
+const Utils = require('../helpers/utils');
 const assert = require('chai').assert;
 const web3 = global.web3;
