+++ test/storage/adddress_manager.js
